TST Uses float32

+++ sklearn/cluster/optics_.py
@@ -233,7 +233,7 @@ if metric=’precomputed’.
         self : instance of OPTICS
             The instance.
         """
-        X = check_array(X, dtype=np.float)
+        X = check_array(X, dtype=np.float32)
 
         if self.cluster_method not in ['dbscan', 'xi']:
             raise ValueError("cluster_method should be one of"
